bugs  squashed, scheduler replaced
